# Conflicts:
#	modules/backend-dossiermgt/backend-dossiermgt-service/src/main/resources/service.properties
#	modules/backend-usermgt/backend-usermgt-service/src/main/resources/service.properties

+++ modules/backend-dossiermgt/backend-dossiermgt-api/src/main/java/org/opencps/dossiermgt/service/ProcessStepLocalServiceWrapper.java
@@ -410,7 +410,7 @@ public class ProcessStepLocalServiceWrapper implements ProcessStepLocalService,
 	public org.opencps.dossiermgt.model.ProcessStep updateProcessStep(
 		long groupId, long processStepId, String stepCode, String stepName,
 		long serviceProcessId, String sequenceNo, String dossierStatus,
-		String dossierSubStatus, int durationCount, String customProcessUrl,
+		String dossierSubStatus, double durationCount, String customProcessUrl,
 		String stepInstruction, String briefNote, boolean editable,
 		String lockState, Integer checkInput,
 		com.liferay.portal.kernel.service.ServiceContext context)
